유튜브 수정

+++ src/main/kotlin/com/formskorea/console/data/model/Media.kt
@@ -9,7 +9,7 @@ data class Media(
     @SerializedName("type") var intType: Int? = null,
     @SerializedName("url") var strURL: String? = null,
     @SerializedName("status") var intStatus: Int? = null,
-    @SerializedName("count1") var intCount1: Int? = null,
-    @SerializedName("count2") var intCount2: Int? = null,
-    @SerializedName("count3") var intCount3: Int? = null
+    @SerializedName("count1") var intCount1: Long? = null,
+    @SerializedName("count2") var intCount2: Long? = null,
+    @SerializedName("count3") var intCount3: Long? = null
 )
